+++ stripe/src/main/java/com/stripe/android/Stripe.kt
@@ -36,6 +36,7 @@ import com.stripe.android.model.StripeModel
 import com.stripe.android.model.Token
 import com.stripe.android.model.TokenParams
 import com.stripe.android.networking.ApiRequest
+import com.stripe.android.networking.StripeApiRepository
 import com.stripe.android.networking.StripeRepository
 import com.stripe.android.view.AuthActivityStarter
 import kotlinx.coroutines.CoroutineScope