+++ stripe/src/test/java/com/stripe/android/networking/FakeAnalyticsRequestExecutor.kt
@@ -1,4 +1,4 @@
-package com.stripe.android
+package com.stripe.android.networking
 
 internal class FakeAnalyticsRequestExecutor : AnalyticsRequestExecutor {
     override fun executeAsync(request: AnalyticsRequest) {}